add test_type_matches to the test machinery
